+++ app/packages/core/src/components/Actions/similar/Helper.tsx
@@ -1,9 +1,9 @@
 import { PopoutSectionTitle, useTheme } from "@fiftyone/components";
-import * as fos from "@fiftyone/state";
 import React from "react";
-import { useRecoilValue } from "recoil";
+
 import { SORT_BY_SIMILARITY } from "../../../utils/links";
 import { ActionOption } from "../Common";
+import { useSimilarityType } from "./utils";
 
 interface Props {
   hasSimilarityKeys: boolean;
@@ -11,20 +11,10 @@ interface Props {
   modal: boolean;
 }
 
-const Warning: React.FunctionComponent<Props> = (props) => {
+const Helper: React.FunctionComponent<Props> = (props) => {
   const theme = useTheme();
-  const hasSelectedSamples = useRecoilValue(fos.hasSelectedSamples);
-  const hasSelectedLabels = useRecoilValue(fos.hasSelectedLabels);
-  const hasSorting = Boolean(useRecoilValue(fos.similarityParameters));
-
-  const isImageSearch =
-    hasSelectedSamples ||
-    (props.isImageSearch && hasSorting) ||
-    (props.modal && hasSelectedLabels);
-
-  const warningText = isImageSearch
-    ? "Search by image similarity"
-    : "Sort by text similarity";
+  const { isImageSearch, modal } = props;
+  const { text } = useSimilarityType({ isImageSearch, modal });
 
   return (
     <>
@@ -39,7 +29,7 @@ const Warning: React.FunctionComponent<Props> = (props) => {
         <ActionOption
           id="sort-by-similarity"
           href={SORT_BY_SIMILARITY}
-          text={warningText}
+          text={text}
           title={"About sorting by similarity"}
           style={{
             background: "unset",
@@ -54,4 +44,4 @@ const Warning: React.FunctionComponent<Props> = (props) => {
   );
 };
 
-export default Warning;
+export default Helper;
